fix(runner): centralize lab workspace labels
